Django/Django-Platzi/coffe-shop/requirements.txt
Django==5.2.1
pillow==11.2.1
djangorestframework==3.16.0
crispy-tailwind==1.0.3
django-environ==0.12.0
psycopg2-binary==2.9.10